old-tests/submission/Google/audio/preload/preload_reflects_no_value.html

<!doctype html>
<html>
 <head>
  <title>audio.preload - reflection test</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#dom-media-preload">spec reference</a></p>
  <audio id="audio">
  </audio>
  <div id="log"></div>
  <script>
test(function() {
 assert_equals(
  document.getElementById("audio").preload,
  "auto",
  "audioElement.preload absence reflects 'none' value");
});
  </script>
 </body>
</html>
